add invariant oracle

+++ tests/evm/echidna_flag/test.sol
@@ -4,11 +4,11 @@ contract Test {
     bool private flag0 = true;
     bool private flag1 = true;
 
-    function set0(int val) public returns (bool) {
+    function set0(int256 val) public returns (bool) {
         if (val % 100 == 0) flag0 = false;
     }
 
-    function set1(int val) public returns (bool) {
+    function set1(int256 val) public returns (bool) {
         if (val % 10 == 0 && !flag0) flag1 = false;
     }
